@@ -3,9 +3,8 @@ import ReactDOM from "react-dom"
 import "./index.css"
 import App from "./components/App/App"
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById("root")
 )
